+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Convpot changelog
 
 ## v1.2.1 - 2017-10-31
-* fix dQ/dV division by zero for rest points 
+* test dQ/dV for "not a number"
+* calculate average current over all data points
+* use average current for half cycle step index
 
 ## v1.2.0 - 2017-10-22
 * verbose option can be given multiple times